requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements/base.in
#
aiofiles==23.2.1
    # via gradio
aiohttp==3.9.3
    # via
    #   langchain
    #   langchain-community
aiosignal==1.3.1
    # via aiohttp
altair==5.2.0
    # via gradio
annotated-types==0.6.0
    # via pydantic
anyio==4.2.0
    # via
    #   httpx
    #   langchain-core
    #   openai
    #   starlette
async-timeout==4.0.3
    # via
    #   aiohttp
    #   langchain
asyncio==3.4.3
    # via -r requirements/base.in
attrs==23.2.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via
    #   pdfminer-six
    #   requests
click==8.1.7
    # via
    #   typer
    #   uvicorn
colorama==0.4.6
    # via typer
contourpy==1.2.0
    # via matplotlib
cryptography==42.0.2
    # via pdfminer-six
cycler==0.12.1
    # via matplotlib
dataclasses-json==0.6.4
    # via
    #   langchain
    #   langchain-community
distro==1.9.0
    # via openai
exceptiongroup==1.2.0
    # via
    #   anyio
    #   pytest
fastapi==0.109.2
    # via gradio
ffmpy==0.3.1
    # via gradio
filelock==3.13.1
    # via huggingface-hub
fonttools==4.48.1
    # via matplotlib
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.2.0
    # via
    #   gradio-client
    #   huggingface-hub
gradio==4.17.0
    # via -r requirements/base.in
gradio-client==0.9.0
    # via gradio
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.2
    # via httpx
httpx==0.26.0
    # via
    #   gradio
    #   gradio-client
    #   openai
huggingface-hub==0.20.3
    # via
    #   gradio
    #   gradio-client
idna==3.6
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-resources==6.1.1
    # via gradio
iniconfig==2.0.0
    # via pytest
jinja2==3.1.3
    # via
    #   altair
    #   gradio
jsonpatch==1.33
    # via
    #   langchain
    #   langchain-core
jsonpointer==2.4
    # via jsonpatch
jsonschema==4.21.1
    # via altair
jsonschema-specifications==2023.12.1
    # via jsonschema
kiwisolver==1.4.5
    # via matplotlib
langchain==0.1.5
    # via -r requirements/base.in
langchain-community==0.0.19
    # via langchain
langchain-core==0.1.21
    # via
    #   langchain
    #   langchain-community
    #   langchain-openai
langchain-openai==0.0.5
    # via -r requirements/base.in
langsmith==0.0.87
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via
    #   gradio
    #   jinja2
marshmallow==3.20.2
    # via dataclasses-json
matplotlib==3.8.2
    # via gradio
mdurl==0.1.2
    # via markdown-it-py
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
numpy==1.26.4
    # via
    #   altair
    #   contourpy
    #   gradio
    #   langchain
    #   langchain-community
    #   langchain-openai
    #   matplotlib
    #   pandas
openai==1.11.1
    # via
    #   -r requirements/base.in
    #   langchain-openai
orjson==3.9.13
    # via gradio
packaging==23.2
    # via
    #   altair
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   langchain-core
    #   marshmallow
    #   matplotlib
    #   pytest
pandas==2.2.0
    # via
    #   altair
    #   gradio
pdfminer-six==20231228
    # via -r requirements/base.in
pillow==10.2.0
    # via
    #   gradio
    #   matplotlib
pluggy==1.4.0
    # via pytest
pycparser==2.21
    # via cffi
pydantic==2.6.1
    # via
    #   fastapi
    #   gradio
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
pydantic-core==2.16.2
    # via pydantic
pydub==0.25.1
    # via gradio
pygments==2.17.2
    # via rich
pyparsing==3.1.1
    # via matplotlib
pytest==8.0.0
    # via -r requirements/base.in
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.0.1
    # via -r requirements/base.in
python-multipart==0.0.7
    # via gradio
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via
    #   gradio
    #   huggingface-hub
    #   langchain
    #   langchain-community
    #   langchain-core
referencing==0.33.0
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2023.12.25
    # via tiktoken
requests==2.31.0
    # via
    #   huggingface-hub
    #   langchain
    #   langchain-community
    #   langchain-core
    #   langsmith
    #   tiktoken
rich==13.7.0
    # via typer
rpds-py==0.17.1
    # via
    #   jsonschema
    #   referencing
ruff==0.2.1
    # via gradio
semantic-version==2.10.0
    # via gradio
shellingham==1.5.4
    # via typer
six==1.16.0
    # via python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
    #   openai
sqlalchemy==2.0.25
    # via
    #   langchain
    #   langchain-community
starlette==0.36.3
    # via fastapi
tenacity==8.2.3
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
tiktoken==0.5.2
    # via langchain-openai
tomli==2.0.1
    # via pytest
tomlkit==0.12.0
    # via gradio
toolz==0.12.1
    # via altair
tqdm==4.66.1
    # via
    #   huggingface-hub
    #   openai
typer[all]==0.9.0
    # via
    #   gradio
    #   typer
typing-extensions==4.9.0
    # via
    #   altair
    #   anyio
    #   fastapi
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typer
    #   typing-inspect
    #   uvicorn
typing-inspect==0.9.0
    # via dataclasses-json
tzdata==2023.4
    # via pandas
urllib3==2.2.0
    # via requests
uvicorn==0.27.0.post1
    # via gradio
websockets==11.0.3
    # via gradio-client
yarl==1.9.4
    # via aiohttp
